feat(console): extract Proposal types to CDN
