+++ src/console/src/storage/state/stable.rs
@@ -1,6 +1,5 @@
 use crate::strategies_impls::cdn::CdnStable;
-use crate::types::state::ProposalId;
-use junobuild_cdn::ProposalAssetKey;
+use junobuild_cdn::{ProposalAssetKey, ProposalId};
 use junobuild_collections::types::core::CollectionKey;
 use junobuild_shared::types::core::Blob;
 use junobuild_storage::types::state::FullPath;
