+++ src/console/src/storage/store.rs
@@ -7,8 +7,8 @@ use crate::storage::state::heap::{
 use crate::store::heap::get_controllers;
 use crate::store::stable::get_proposal;
 use crate::strategies_impls::storage::StorageState;
-use crate::types::state::ProposalId;
 use candid::Principal;
+use junobuild_cdn::ProposalId;
 use junobuild_collections::types::core::CollectionKey;
 use junobuild_collections::types::rules::Memory;
 use junobuild_shared::list::list_values;
