+++ src/console/src/store/stable.rs
@@ -2,12 +2,13 @@ use crate::constants::E8S_PER_ICP;
 use crate::memory::STATE;
 use crate::types::ledger::{Payment, PaymentStatus};
 use crate::types::state::{
-    MissionControl, MissionControls, MissionControlsStable, Payments, PaymentsStable, Proposal,
-    ProposalId, ProposalKey, ProposalsStable, StableState,
+    MissionControl, MissionControls, MissionControlsStable, Payments, PaymentsStable,
+    ProposalsStable, StableState,
 };
 use ic_cdk::api::time;
 use ic_ledger_types::BlockIndex;
 use ic_ledger_types::Tokens;
+use junobuild_cdn::{Proposal, ProposalId, ProposalKey};
 use junobuild_shared::types::state::MissionControlId;
 use junobuild_shared::types::state::UserId;
 use junobuild_shared::utils::principal_equal;
