+++ src/libs/cdn/src/lib.rs
@@ -1,9 +1,10 @@
+mod proposals;
 mod storage;
 pub mod strategies;
-mod types;
 
+pub use proposals::state::types::*;
+pub use proposals::types::*;
 pub use storage::certified_assets::init_certified_assets;
 pub use storage::state::heap;
 pub use storage::state::stable;
 pub use storage::state::types::*;
-pub use types::state::*;
