+++ src/libs/cdn/src/proposals/mod.rs
@@ -1,0 +1,3 @@
+mod impls;
+pub mod state;
+pub mod types;
