+++ src/libs/cdn/src/proposals/types.rs
@@ -1,0 +1,9 @@
+#[derive(Debug)]
+pub enum CommitProposalError {
+    ProposalNotFound(String),
+    ProposalNotOpen(String),
+    InvalidSha256(String),
+    InvalidType(String),
+    CommitAssetsIssue(String),
+    PostCommitAssetsIssue(String),
+}
