+++ src/libs/cdn/src/storage/state/types.rs
@@ -1,4 +1,4 @@
-use crate::types::state::ProposalId;
+use crate::proposals::state::types::ProposalId;
 use candid::{CandidType, Deserialize};
 use ic_stable_structures::StableBTreeMap;
 use junobuild_collections::types::core::CollectionKey;
