--- src/libs/cdn/src/types.rs
@@ -1,3 +1,0 @@
-pub mod state {
-    pub type ProposalId = u128;
-}
